requirements.txt
qdrant-client
beautifulsoup4
langchain
httpx
sentence-transformers
typer
python-dotenv
